processor typescript conversion started
