--- unfetter-discover-processor/.eslintrc.json
@@ -1,42 +1,0 @@
-{
-  "extends": "airbnb-base",
-  "plugins": [
-    "import"
-  ],
-  "rules": {
-    // ignore the dangle rule, mongo used _id
-    "no-underscore-dangle": 0,
-    "comma-dangle": 0,
-    "no-console": 0,
-    "consistent-return": 0,
-    "max-len": 0,
-    "no-restricted-syntax": 0,
-    "no-var": 0,
-    "indent": [
-      "warn",
-      4
-    ],
-    "arrow-parens": [
-      "error",
-      "as-needed"
-    ],
-    "no-void": 0,
-    "import/no-unresolved": "off",
-    "no-shadow": ["warn", { "builtinGlobals": false, "hoist": "functions", "allow": [] }],
-    "global-require":"off",
-    "prefer-destructuring": [1, {
-      "array": true,
-      "object": false
-    }, {
-      "enforceForRenamedProperties": false
-    }],
-    "no-use-before-define": "off",
-    "new-cap":"off",
-    "no-unused-vars":"warn",
-    "no-param-reassign":"off",
-    "camelcase":"warn"
-  },
-  "env": {
-    "jasmine": true
-  }
-}
