+++ unfetter-discover-processor/package.json
@@ -2,10 +2,18 @@
   "name": "unfetter-discover-processor",
   "version": "0.3.6",
   "description": "Records Processor for Unfetter Discover",
-  "private": true,
-  "directories": {
-    "doc": "doc",
-    "test": "tests"
+  "scripts": {
+    "checkup": "npm run clean && npm run build && npm run lint && npm run coverage",
+    "coverage": "nyc --extension .ts --reporter=lcov --reporter=text-lcov npm test && nyc report",
+    "clean": "rimraf dist/ tsdocs/ coverage/ .nyc_output/",
+    "start": "npm run build && npm run watch",
+    "build": "npm run build-ts",
+    "start:dev": "nodemon dist/app.js",
+    "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run start:dev\"",
+    "build-ts": "tsc",
+    "watch-ts": "tsc -w",
+    "lint": "tsc -v && tslint src/**/*.ts",
+    "test": "jasmine-ts --config=jasmine.json"
   },
   "author": "Unfetter",
   "license": "MIT",
@@ -16,11 +24,18 @@
     "yargs": "^8.0.2"
   },
   "devDependencies": {
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb-base": "^11.2.0",
-    "eslint-plugin-import": "^2.3.0"
-  },
-  "scripts": {
-    "lint": "eslint ./*.js"
+    "@types/mongoose": "^5.0.10",
+    "@types/node": "^9.6.6",
+    "@types/node-fetch": "^1.6.8",
+    "@types/yargs": "^11.0.0",
+    "concurrently": "^3.5.1",
+    "jasmine": "^3.1.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "jasmine-ts": "^0.2.1",
+    "nodemon": "^1.17.3",
+    "nyc": "^11.7.1",
+    "rimraf": "^2.6.2",
+    "tslint": "^5.9.1",
+    "typescript": "^2.8.3"
   }
 }
