+++ unfetter-discover-processor/src/vendor.d.ts
@@ -1,0 +1,4 @@
+declare module 'https-proxy-agent' {
+    const x: any;
+    export = x;
+}
